Set lessons routes
